--- App.js
@@ -1,1288 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-
-const styleCard = {
-  backgroundColor: "#f0f0f0"
-}
-
-const rseObj = [
-  {
-      "info": {
-          "id": "634125",
-          "name": "Subway",
-          "cloudinaryImageId": "63178e3e64d503a479f2a2048a474552",
-          "locality": "1st Phase",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹350 for two",
-          "cuisines": [
-              "Salads",
-              "Snacks",
-              "Desserts",
-              "Beverages"
-          ],
-          "avgRating": 3.8,
-          "parentId": "2",
-          "avgRatingString": "3.8",
-          "totalRatingsString": "100+",
-          "sla": {
-              "deliveryTime": 22,
-              "lastMileTravel": 3.5,
-              "serviceability": "SERVICEABLE",
-              "slaString": "22 mins",
-              "lastMileTravelString": "3.5 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 02:00:00",
-              "opened": true
-          },
-          "badges": {
-              "textExtendedBadges": [
-                  {
-                      "iconId": "guiltfree/GF_Logo_android_3x",
-                      "shortDescription": "options available",
-                      "fontColor": "#7E808C"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "",
-                                  "fontColor": "#7E808C",
-                                  "iconId": "guiltfree/GF_Logo_android_3x",
-                                  "shortDescription": "options available"
-                              }
-                          }
-                      ]
-                  }
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹100 OFF",
-              "subHeader": "ABOVE ₹449",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/subway-1st-phase-mayur-vihar-delhi-634125",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "462638",
-          "name": "Domino's Pizza",
-          "cloudinaryImageId": "zh21rugnoq9lkijtegg2",
-          "locality": "2nd Phase",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹400 for two",
-          "cuisines": [
-              "Pizzas",
-              "Italian",
-              "Pastas",
-              "Desserts"
-          ],
-          "avgRating": 4.4,
-          "parentId": "2456",
-          "avgRatingString": "4.4",
-          "totalRatingsString": "1K+",
-          "sla": {
-              "deliveryTime": 30,
-              "serviceability": "SERVICEABLE",
-              "slaString": "30 mins",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 02:59:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹150 OFF",
-              "subHeader": "ABOVE ₹299",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/dominos-pizza-2nd-phase-mayur-vihar-delhi-462638",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "623102",
-          "name": "Khadak Singh Da Dhaba",
-          "cloudinaryImageId": "8144bf37947c161953f656308d1bfc30",
-          "locality": "Pandav Nagar",
-          "areaName": "Mayur Vihar Phase 1",
-          "costForTwo": "₹300 for two",
-          "cuisines": [
-              "North Indian",
-              "Chinese",
-              "Mughlai",
-              "Tandoor"
-          ],
-          "avgRating": 4,
-          "parentId": "13170",
-          "avgRatingString": "4.0",
-          "totalRatingsString": "1K+",
-          "sla": {
-              "deliveryTime": 19,
-              "lastMileTravel": 2.6,
-              "serviceability": "SERVICEABLE",
-              "slaString": "19 mins",
-              "lastMileTravelString": "2.6 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 01:45:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "50% OFF",
-              "subHeader": "UPTO ₹100"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/khadak-singh-da-dhaba-pandav-nagar-mayur-vihar-phase-1-delhi-623102",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "681857",
-          "name": "Louis Burger",
-          "cloudinaryImageId": "a3535141ce7f8db939e999c2a7e06a0c",
-          "locality": "School Block",
-          "areaName": "Shakarpur",
-          "costForTwo": "₹600 for two",
-          "cuisines": [
-              "Burgers",
-              "American",
-              "Fast Food"
-          ],
-          "avgRating": 4.3,
-          "parentId": "22485",
-          "avgRatingString": "4.3",
-          "totalRatingsString": "500+",
-          "sla": {
-              "deliveryTime": 24,
-              "lastMileTravel": 3.9,
-              "serviceability": "SERVICEABLE",
-              "slaString": "24 mins",
-              "lastMileTravelString": "3.9 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 04:00:00",
-              "opened": true
-          },
-          "badges": {
-              "imageBadges": [
-                  {
-                      "imageId": "newg.png",
-                      "description": "Gourmet"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "Gourmet",
-                                  "imageId": "newg.png"
-                              }
-                          }
-                      ]
-                  },
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹125 OFF",
-              "subHeader": "ABOVE ₹249",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/louis-burger-school-block-shakarpur-delhi-681857",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "720360",
-          "name": "Wendy's Burgers",
-          "cloudinaryImageId": "f1aa621021a2826088089b89842d4e7c",
-          "locality": "Laxmi Nagar",
-          "areaName": "Shakarpur",
-          "costForTwo": "₹200 for two",
-          "cuisines": [
-              "Burgers",
-              "American",
-              "Fast Food",
-              "Snacks",
-              "Beverages"
-          ],
-          "avgRating": 4.2,
-          "parentId": "972",
-          "avgRatingString": "4.2",
-          "totalRatingsString": "500+",
-          "sla": {
-              "deliveryTime": 25,
-              "lastMileTravel": 4.3,
-              "serviceability": "SERVICEABLE",
-              "slaString": "25 mins",
-              "lastMileTravelString": "4.3 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-10 23:59:00",
-              "opened": true
-          },
-          "badges": {
-              "textExtendedBadges": [
-                  {
-                      "iconId": "guiltfree/GF_Logo_android_3x",
-                      "shortDescription": "options available",
-                      "fontColor": "#7E808C"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "",
-                                  "fontColor": "#7E808C",
-                                  "iconId": "guiltfree/GF_Logo_android_3x",
-                                  "shortDescription": "options available"
-                              }
-                          }
-                      ]
-                  }
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "ITEMS",
-              "subHeader": "AT ₹149"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/wendys-burgers-laxmi-nagar-shakarpur-delhi-720360",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "685709",
-          "name": "MOJO Pizza - 2X Toppings",
-          "cloudinaryImageId": "59b64f9bb67234987652437f21d379c2",
-          "locality": "Patparganj",
-          "areaName": "Pandav Nagar Main Road",
-          "costForTwo": "₹250 for two",
-          "cuisines": [
-              "Pizzas",
-              "Italian",
-              "Fast Food",
-              "Desserts"
-          ],
-          "avgRating": 4.3,
-          "parentId": "11329",
-          "avgRatingString": "4.3",
-          "totalRatingsString": "100+",
-          "sla": {
-              "deliveryTime": 33,
-              "lastMileTravel": 3,
-              "serviceability": "SERVICEABLE",
-              "slaString": "28-38 mins",
-              "lastMileTravelString": "3.0 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 02:00:00",
-              "opened": true
-          },
-          "badges": {
-              "textExtendedBadges": [
-                  {
-                      "iconId": "guiltfree/GF_Logo_android_3x",
-                      "shortDescription": "options available",
-                      "fontColor": "#7E808C"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "",
-                                  "fontColor": "#7E808C",
-                                  "iconId": "guiltfree/GF_Logo_android_3x",
-                                  "shortDescription": "options available"
-                              }
-                          }
-                      ]
-                  }
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹100 OFF",
-              "subHeader": "ABOVE ₹249",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/mojo-pizza-2x-toppings-patparganj-pandav-nagar-main-road-delhi-685709",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "253981",
-          "name": "McDonald's",
-          "cloudinaryImageId": "ecd32b841549ce6bfd993260a2086c1f",
-          "locality": "V3S Mall",
-          "areaName": "Laxmi Nagar",
-          "costForTwo": "₹400 for two",
-          "cuisines": [
-              "American"
-          ],
-          "avgRating": 4.4,
-          "parentId": "630",
-          "avgRatingString": "4.4",
-          "totalRatingsString": "10K+",
-          "sla": {
-              "deliveryTime": 32,
-              "lastMileTravel": 4.3,
-              "serviceability": "SERVICEABLE",
-              "slaString": "32 mins",
-              "lastMileTravelString": "4.3 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-10 23:45:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "aggregatedDiscountInfoV2": {},
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/mcdonalds-v3s-mall-laxmi-nagar-delhi-253981",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "438435",
-          "name": "Burger King",
-          "cloudinaryImageId": "e33e1d3ba7d6b2bb0d45e1001b731fcf",
-          "locality": "Parsvnath Mall",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹350 for two",
-          "cuisines": [
-              "Burgers",
-              "American"
-          ],
-          "avgRating": 4.3,
-          "parentId": "166",
-          "avgRatingString": "4.3",
-          "totalRatingsString": "1K+",
-          "sla": {
-              "deliveryTime": 47,
-              "lastMileTravel": 3.6,
-              "serviceability": "SERVICEABLE",
-              "slaString": "47 mins",
-              "lastMileTravelString": "3.6 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 05:00:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "60% OFF",
-              "subHeader": "UPTO ₹120"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/burger-king-parsvnath-mall-mayur-vihar-delhi-438435",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "516395",
-          "name": "Hilton restaurant",
-          "cloudinaryImageId": "a6ylbndocvy5i64vjz31",
-          "locality": "Mandawali",
-          "areaName": "Pandav Nagar",
-          "costForTwo": "₹450 for two",
-          "cuisines": [
-              "Biryani",
-              "Fast Food",
-              "Snacks",
-              "Ice Cream"
-          ],
-          "avgRating": 3.2,
-          "parentId": "19233",
-          "avgRatingString": "3.2",
-          "totalRatingsString": "20+",
-          "sla": {
-              "deliveryTime": 29,
-              "lastMileTravel": 3,
-              "serviceability": "SERVICEABLE",
-              "slaString": "29 mins",
-              "lastMileTravelString": "3.0 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-10 23:59:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "aggregatedDiscountInfoV2": {},
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/hilton-restaurant-mandawali-pandav-nagar-delhi-516395",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "655722",
-          "name": "Veer Ji Malai Chaap Wale",
-          "cloudinaryImageId": "b139cc0e94645b50b1bca2bd8dced0cd",
-          "locality": "Lsc Market",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹300 for two",
-          "cuisines": [
-              "North Indian",
-              "Biryani"
-          ],
-          "avgRating": 3.7,
-          "parentId": "12056",
-          "avgRatingString": "3.7",
-          "totalRatingsString": "100+",
-          "sla": {
-              "deliveryTime": 16,
-              "lastMileTravel": 1.5,
-              "serviceability": "SERVICEABLE",
-              "slaString": "16 mins",
-              "lastMileTravelString": "1.5 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 02:00:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "aggregatedDiscountInfoV2": {},
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/veer-ji-malai-chaap-wale-lsc-market-mayur-vihar-delhi-655722",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "636891",
-          "name": "Big Bowl",
-          "cloudinaryImageId": "c99751d54e4e1847186c62b3309c1327",
-          "locality": "1st Phase",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹250 for two",
-          "cuisines": [
-              "North Indian",
-              "Chinese",
-              "Tibetan",
-              "Desserts"
-          ],
-          "avgRating": 4.2,
-          "parentId": "434792",
-          "avgRatingString": "4.2",
-          "totalRatingsString": "500+",
-          "sla": {
-              "deliveryTime": 22,
-              "lastMileTravel": 3.5,
-              "serviceability": "SERVICEABLE",
-              "slaString": "22 mins",
-              "lastMileTravelString": "3.5 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 03:00:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "ITEMS",
-              "subHeader": "AT ₹169"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/big-bowl-1st-phase-mayur-vihar-delhi-636891",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "769955",
-          "name": "Olio - The Wood Fired Pizzeria",
-          "cloudinaryImageId": "4b44c7921b1b6073a4ffa58b38f8def1",
-          "locality": "Near Reliance Fresh",
-          "areaName": "East Vinod Nagar",
-          "costForTwo": "₹300 for two",
-          "cuisines": [
-              "Pizzas",
-              "Pastas",
-              "Fast Food",
-              "Snacks",
-              "Beverages"
-          ],
-          "avgRating": 4.4,
-          "parentId": "11633",
-          "avgRatingString": "4.4",
-          "totalRatingsString": "100+",
-          "sla": {
-              "deliveryTime": 27,
-              "lastMileTravel": 2.6,
-              "serviceability": "SERVICEABLE",
-              "slaString": "27 mins",
-              "lastMileTravelString": "2.6 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-10 23:59:00",
-              "opened": true
-          },
-          "badges": {
-              "imageBadges": [
-                  {
-                      "imageId": "newg.png",
-                      "description": "Gourmet"
-                  }
-              ],
-              "textExtendedBadges": [
-                  {
-                      "iconId": "guiltfree/GF_Logo_android_3x",
-                      "shortDescription": "options available",
-                      "fontColor": "#7E808C"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "Gourmet",
-                                  "imageId": "newg.png"
-                              }
-                          }
-                      ]
-                  },
-                  "textBased": {},
-                  "textExtendedBadges": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "",
-                                  "fontColor": "#7E808C",
-                                  "iconId": "guiltfree/GF_Logo_android_3x",
-                                  "shortDescription": "options available"
-                              }
-                          }
-                      ]
-                  }
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹125 OFF",
-              "subHeader": "ABOVE ₹249",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "isNewlyOnboarded": true,
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/olio-the-wood-fired-pizzeria-near-reliance-fresh-east-vinod-nagar-delhi-769955",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "399071",
-          "name": "Cheesecake & Co.",
-          "cloudinaryImageId": "b318c0b4bc2169550145ace1d6e791a2",
-          "locality": "Village Naya Bans",
-          "areaName": "Sector 15",
-          "costForTwo": "₹350 for two",
-          "cuisines": [
-              "Bakery",
-              "Desserts"
-          ],
-          "avgRating": 4.6,
-          "veg": true,
-          "parentId": "387417",
-          "avgRatingString": "4.6",
-          "totalRatingsString": "1K+",
-          "sla": {
-              "deliveryTime": 23,
-              "lastMileTravel": 5,
-              "serviceability": "SERVICEABLE",
-              "slaString": "23 mins",
-              "lastMileTravelString": "5.0 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 02:00:00",
-              "opened": true
-          },
-          "badges": {
-              "imageBadges": [
-                  {
-                      "imageId": "newg.png",
-                      "description": "Gourmet"
-                  },
-                  {
-                      "imageId": "v1695133679/badges/Pure_Veg111.png",
-                      "description": "pureveg"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "Gourmet",
-                                  "imageId": "newg.png"
-                              }
-                          },
-                          {
-                              "attributes": {
-                                  "description": "pureveg",
-                                  "imageId": "v1695133679/badges/Pure_Veg111.png"
-                              }
-                          }
-                      ]
-                  },
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "₹75 OFF",
-              "subHeader": "ABOVE ₹199",
-              "discountTag": "FLAT DEAL"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/cheesecake-and-co-village-naya-bans-sector-15-delhi-399071",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "744443",
-          "name": "Grameen Kulfi",
-          "cloudinaryImageId": "930b886678eb9de74b5a3150bdd0f405",
-          "locality": "Pandav nagar",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹120 for two",
-          "cuisines": [
-              "Ice Cream",
-              "Desserts"
-          ],
-          "avgRating": 4.1,
-          "veg": true,
-          "parentId": "12175",
-          "avgRatingString": "4.1",
-          "totalRatingsString": "10+",
-          "sla": {
-              "deliveryTime": 16,
-              "lastMileTravel": 1.5,
-              "serviceability": "SERVICEABLE",
-              "slaString": "16 mins",
-              "lastMileTravelString": "1.5 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 00:00:00",
-              "opened": true
-          },
-          "badges": {},
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {},
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "60% OFF",
-              "subHeader": "UPTO ₹120"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/grameen-kulfi-pandav-nagar-mayur-vihar-delhi-744443",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  },
-  {
-      "info": {
-          "id": "673902",
-          "name": "Goila Butter Chicken",
-          "cloudinaryImageId": "5e19832da032dd69547565e27104706f",
-          "locality": "Block-159",
-          "areaName": "Mayur Vihar",
-          "costForTwo": "₹600 for two",
-          "cuisines": [
-              "North Indian",
-              "Biryani",
-              "Kebabs",
-              "Desserts"
-          ],
-          "avgRating": 4.1,
-          "parentId": "322587",
-          "avgRatingString": "4.1",
-          "totalRatingsString": "100+",
-          "sla": {
-              "deliveryTime": 35,
-              "lastMileTravel": 4.1,
-              "serviceability": "SERVICEABLE",
-              "slaString": "35 mins",
-              "lastMileTravelString": "4.1 km",
-              "iconType": "ICON_TYPE_EMPTY"
-          },
-          "availability": {
-              "nextCloseTime": "2024-02-11 00:00:00",
-              "opened": true
-          },
-          "badges": {
-              "imageBadges": [
-                  {
-                      "imageId": "newg.png",
-                      "description": "Gourmet"
-                  }
-              ]
-          },
-          "isOpen": true,
-          "type": "F",
-          "badgesV2": {
-              "entityBadges": {
-                  "imageBased": {
-                      "badgeObject": [
-                          {
-                              "attributes": {
-                                  "description": "Gourmet",
-                                  "imageId": "newg.png"
-                              }
-                          }
-                      ]
-                  },
-                  "textBased": {},
-                  "textExtendedBadges": {}
-              }
-          },
-          "aggregatedDiscountInfoV3": {
-              "header": "ITEMS",
-              "subHeader": "AT ₹199"
-          },
-          "orderabilityCommunication": {
-              "title": {},
-              "subTitle": {},
-              "message": {},
-              "customIcon": {}
-          },
-          "differentiatedUi": {
-              "displayType": "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-              "differentiatedUiMediaDetails": {
-                  "mediaType": "ADS_MEDIA_ENUM_IMAGE",
-                  "lottie": {},
-                  "video": {}
-              }
-          },
-          "reviewsSummary": {},
-          "displayType": "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-          "restaurantOfferPresentationInfo": {}
-      },
-      "analytics": {
-          "context": "seo-data-508efa18-ebe2-48b6-b230-f0a0c367638e"
-      },
-      "cta": {
-          "link": "https://www.swiggy.com/restaurants/goila-butter-chicken-block-159-mayur-vihar-delhi-673902",
-          "text": "RESTAURANT_MENU",
-          "type": "WEBLINK"
-      },
-      "widgetId": "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo"
-  }
-];
-
-const RestaurantCrad = ({resData}) => {
-  const { cloudinaryImageId, name, cuisines, avgRating, sla : { slaString },  costForTwo } = resData?.info;
-  return (
-    <div className="res-card" style={styleCard}>
-      <img className="res-card-img" src= {`https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/${ cloudinaryImageId}`} />
-       <h3> { name } </h3>
-       <h4 className="cuisine"> { cuisines.join(',') } </h4>
-       <h4>{ avgRating } </h4>
-       <h4> { slaString }  </h4>
-       <h4> { costForTwo }  </h4>
-    </div>
-  )
-}
-
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="search">Search</div>
-      <div className="res-container">
-       
-
-        {
-          rseObj?.map((res) => (
-            <RestaurantCrad resData = {res} key={res?.info?.id}  />
-          ))
-        }
-        
-      </div>
-    </div>
-  )
-}
-
-const Header = () => {
-  return (
-    <div className="header">
-      <div className="logo-container">
-        <img className="logo" src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdaMYtVi9_tfNcpsbGGseU6ehYgV9UeU3h7A&usqp=CAU"></img>
-      </div>
-      <div className="nav-items">
-      <ul>
-        <li>Home</li>
-        <li>About Us</li>
-        <li>Contact Us</li>
-        <li>Cart</li>
-      </ul>
-      </div>
-    </div>
-  )
-}
-
-const AppLayout = () => {
-  return (
-    <div className="app">
-      <Header/>
-      <Body/>
-    </div>
-  )
-}
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<AppLayout />);
