+++ index.css
@@ -39,3 +39,6 @@
     text-overflow: ellipsis;
     overflow-wrap: break-word;
 }
+.filter {
+    margin: 10px;
+}
